+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Sequence` class: make `__repr__` more friendly for long sequences.
 - `Sequence` class: add support for `str` in `__eq__`.
 - `Sequence` class: add a `__hash__` method.
+- `Sequence` class: rename the `id` property to `accession`.
 
 ## [1.0.0] - 2022-07-12
 ### Added